[columnar] [Bug] Sequential scan parallel execution OOM
When assigning next stripe to be read, workers need to scan complete
stripe metadata table and reach stripe which need to be read next.
During this fetch and compare of stripe metadata rows memory is never
released for stripes that needs to be skipped.
This should be addressed by releasing memory that is not needed.