[ci] turn off fail-fast for pg version matrix
In most cases, the build is most likely to fail nearly simultaneously,
so this will let each build run until it errors rather than aborting
just before it was likely to fail, confirming the problem is not
version specific.

In cases where the problem is version specific, this will inform us if
the problem is actually specific to a particular version of Postgres.