Revert "[columnar] Adjust plan for parallel workers (#98)" (#112)
This reverts commit e7038b96bc270e4341f9fbf33abbf29439f60521.

After review of benchmarks, we found the performance impact of this
change to be significant. By adjusting the autovacuum settings to
`analyze` more frequently, and/or by running `analyze` after inserting a
large amount of data (a common best practice for Postgres), users should
be able to avoid the issue which we were fixing in #98.

In other words, making optimized databases slower so unoptimized
databases can run faster wasn't a good tradeoff. Best of both worlds
would be even better, but for now we'll take the faster result.