Fix memory leak when executing vectorized quals (#242)
* Fix memory leak when executing vectorized quals

Currently, a VectorColumn structure is created when executing vectorized
quals for each tuple on the ExecutorState memory context.  However, it
will be freed only until execution finishes.  This commit changes the
memory context to a tuple memory context.

* Fix memory leak in ReadStripeNextVector()

In ReadStripeNextVector(), the memory of columnValueOffset is allocated
in ExecutorState memory content, it will be freed until execution is
finished, so call pfree() to explicitly release the memory to avoid
memory growing up.